+++ frontend/src/scenes/funnels/FunnelBarChart.tsx
@@ -6,7 +6,7 @@ import { AvailableFeature, ChartParams, FunnelStepWithConversionMetrics } from '
 import { LemonRow } from 'lib/components/LemonRow'
 import { Lettermark, LettermarkColor } from 'lib/components/Lettermark/Lettermark'
 import { EntityFilterInfo } from 'lib/components/EntityFilterInfo'
-import { getActionFilterFromFunnelStep } from 'scenes/insights/InsightTabs/FunnelTab/funnelStepTableUtils'
+import { getActionFilterFromFunnelStep } from 'scenes/insights/Views/Funnels/funnelStepTableUtils'
 import { IconSchedule, IconTrendingFlat, IconTrendingFlatDown } from 'lib/components/icons'
 import { humanFriendlyDuration, percentage, pluralize } from 'lib/utils'
 import { ValueInspectorButton } from './ValueInspectorButton'
