+++ frontend/src/scenes/funnels/useFunnelTooltip.tsx
@@ -6,7 +6,7 @@ import { FunnelStepWithConversionMetrics } from '~/types'
 import { LemonRow } from 'lib/components/LemonRow'
 import { Lettermark, LettermarkColor } from 'lib/components/Lettermark/Lettermark'
 import { EntityFilterInfo } from 'lib/components/EntityFilterInfo'
-import { getActionFilterFromFunnelStep } from 'scenes/insights/InsightTabs/FunnelTab/funnelStepTableUtils'
+import { getActionFilterFromFunnelStep } from 'scenes/insights/Views/Funnels/funnelStepTableUtils'
 import { humanFriendlyDuration, humanFriendlyNumber, percentage } from 'lib/utils'
 import { ensureTooltipElement } from 'scenes/insights/Views/LineGraph/LineGraph'
 import { LemonDivider } from 'lib/components/LemonDivider'
