+++ frontend/src/scenes/insights/EditorFilters/EFFunnelsAdvanced.tsx
@@ -2,10 +2,10 @@ import React from 'react'
 import { useValues, useActions, useMountedLogic } from 'kea'
 import { funnelLogic } from 'scenes/funnels/funnelLogic'
 import { FunnelStepReference, StepOrderValue, EditorFilterProps } from '~/types'
-import { FunnelStepOrderPicker } from '../InsightTabs/FunnelTab/FunnelStepOrderPicker'
+import { FunnelStepOrderPicker } from '../Views/Funnels/FunnelStepOrderPicker'
 import { FunnelExclusionsFilter } from '../Filters/FunnelExclusionsFilter'
 import { FunnelStepReferencePicker } from '../Filters/FunnelStepReferencePicker'
-import { funnelCommandLogic } from '../InsightTabs/FunnelTab/funnelCommandLogic'
+import { funnelCommandLogic } from '../Views/Funnels/funnelCommandLogic'
 import { LemonButton } from 'lib/components/LemonButton'
 import { EditorFilterItemTitle } from './EditorFilterItemTitle'
