+++ frontend/src/scenes/insights/InsightContainer.tsx
@@ -29,7 +29,7 @@ import { insightSceneLogic } from 'scenes/insights/insightSceneLogic'
 import { Tooltip } from 'lib/components/Tooltip'
 import { LemonButton } from 'lib/components/LemonButton'
 import { IconExport } from 'lib/components/icons'
-import { FunnelStepsTable } from './InsightTabs/FunnelTab/FunnelStepsTable'
+import { FunnelStepsTable } from './Views/Funnels/FunnelStepsTable'
 import { Animation } from 'lib/components/Animation/Animation'
 import { AnimationType } from 'lib/animations/animations'
 import { FunnelCorrelation } from './Views/FunnelCorrelation'
