+++ frontend/src/scenes/insights/InsightDisplayConfig.tsx
@@ -8,12 +8,12 @@ import { FilterType, FunnelVizType, ItemMode, InsightType, ChartDisplayType } fr
 import { CalendarOutlined } from '@ant-design/icons'
 import { InsightDateFilter } from './Filters/InsightDateFilter'
 import { RetentionDatePicker } from './RetentionDatePicker'
-import { FunnelDisplayLayoutPicker } from './InsightTabs/FunnelTab/FunnelDisplayLayoutPicker'
-import { PathStepPicker } from './InsightTabs/PathTab/PathStepPicker'
+import { FunnelDisplayLayoutPicker } from './Views/Funnels/FunnelDisplayLayoutPicker'
+import { PathStepPicker } from './Views/Paths/PathStepPicker'
 import { ReferencePicker as RetentionReferencePicker } from './Filters/ReferencePicker'
 import { Tooltip } from 'antd'
 import { InfoCircleOutlined } from '@ant-design/icons'
-import { FunnelBinsPicker } from './InsightTabs/FunnelTab/FunnelBinsPicker'
+import { FunnelBinsPicker } from './Views/Funnels/FunnelBinsPicker'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
 import { useValues } from 'kea'
 import { FEATURE_FLAGS } from 'lib/constants'
