--- frontend/src/scenes/insights/InsightTabs/TrendTab/index.ts
@@ -1,1 +1,0 @@
-export * from './TrendTab'
