--- frontend/src/scenes/insights/InsightTabs/index.ts
@@ -1,4 +1,0 @@
-export * from './RetentionTab'
-export * from './TrendTab'
-export * from './PathTab/PathTab'
-export * from './FunnelTab'
