+++ frontend/src/scenes/insights/InsightsNav.tsx
@@ -5,7 +5,7 @@ import { InsightType } from '~/types'
 import { insightLogic } from './insightLogic'
 import { Tooltip } from 'lib/components/Tooltip'
 import clsx from 'clsx'
-import { FunnelsCue } from './InsightTabs/TrendTab/FunnelsCue'
+import { FunnelsCue } from './Views/Trends/FunnelsCue'
 import { INSIGHT_TYPES_METADATA } from 'scenes/saved-insights/SavedInsights'
 import { Link } from 'lib/components/Link'
 import { urls } from 'scenes/urls'
