+++ frontend/src/scenes/insights/Views/Funnels/CorrelationMatrix.scss
