+++ frontend/src/scenes/insights/Views/Funnels/FunnelBinsPicker.tsx
