+++ frontend/src/scenes/insights/Views/Funnels/FunnelCorrelationTable.scss
