+++ frontend/src/scenes/insights/Views/Funnels/FunnelDisplayLayoutPicker.tsx
