+++ frontend/src/scenes/insights/Views/Funnels/FunnelPropertyCorrelationTable.tsx
