+++ frontend/src/scenes/insights/Views/Funnels/FunnelStepsPicker.tsx
