+++ frontend/src/scenes/insights/Views/Funnels/FunnelStepsTable.tsx
