+++ frontend/src/scenes/insights/Views/Funnels/funnelCommandLogic.ts
