+++ frontend/src/scenes/insights/Views/Funnels/funnelStepTableUtils.tsx
