+++ frontend/src/scenes/insights/Views/Paths/PathStepPicker.tsx
