+++ frontend/src/scenes/insights/Views/Trends/FunnelsCue.scss
