+++ frontend/src/scenes/insights/Views/Trends/FunnelsCue.tsx
@@ -6,7 +6,7 @@ import React from 'react'
 import clsx from 'clsx'
 import './FunnelsCue.scss'
 import { Button } from 'antd'
-import { funnelsCueLogic } from 'scenes/insights/InsightTabs/TrendTab/funnelsCueLogic'
+import { funnelsCueLogic } from 'scenes/insights/Views/Trends/funnelsCueLogic'
 import { insightLogic } from 'scenes/insights/insightLogic'
 import { urls } from 'scenes/urls'
 import { InsightType } from '~/types'
