+++ frontend/src/scenes/insights/Views/Trends/funnelsCueLogic.tsx
